Revert "Addressed Fortran  issues with promoted integers and reals via compil…" (#4220)
This reverts commit 06c42ff0385b34317ecdc06b6aaf6e435fbc07ca.

+++ fortran/src/H5Sf.c
@@ -55,7 +55,7 @@ h5screate_simple_c(int_f *rank, hsize_t_f *dims, hsize_t_f *maxdims, hid_t_f *sp
         c_maxdims[i] = maxdims[*rank - i - 1];
     } /* end for */
 
-    c_space_id = H5Screate_simple((int)*rank, c_dims, c_maxdims);
+    c_space_id = H5Screate_simple(*rank, c_dims, c_maxdims);
     if (c_space_id < 0)
         HGOTO_DONE(FAIL);
